MDEV-24248: my_print_defaults is not taking all the values when using -e option which is called from mysql.server (extra_args).
Analysis: When get_default_options() is called again when reading options
from config file (my_load_defaults()), my_defaults_extra_file is reset when
my_defaults* is initialized to 0.
This is not a problem when we first call get_defaults_options()
in main() because we dont have values in my_defaults* yet,
they are assigned during handle_options(). But we only call handle_option()
once, so the values are not "reassigned" after the second call
to get_default_options()
Fix: Save old values of my_defaults* if we are using my_print_defaults and
reassign them later.